requirements.txt
faster-whisper
SQLAlchemy
torch
tqdm
bilibili-api-python
qrcode
Pillow
crewai
litellm
python-dotenv
fastapi
email-validator
fastapi-sso